# docs/swagger.yaml
basePath: /api
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  helpers.GenericDeleteSuccessResponse:
    properties:
      code:
        example: 204
        type: integer
    type: object
  helpers.GenericErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: parsing id failed
        type: string
    type: object
  model.Alarm:
    properties:
      active:
        type: boolean
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      muted:
        type: boolean
      name:
        type: string
      notifications:
        items:
          $ref: '#/definitions/model.Notification'
        type: array
      targetId:
        type: integer
      threshold:
        type: number
      thresholdField:
        $ref: '#/definitions/model.AlarmThresholdField'
      type:
        $ref: '#/definitions/model.AlarmType'
      updatedAt:
        type: string
    type: object
  model.AlarmThresholdField:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - Latency
    - thresholdFieldMax
  model.AlarmType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Unavailable
    - Threshold
    - alarmTypeMax
  model.Checker:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      heartbeats:
        items:
          $ref: '#/definitions/model.Heartbeat'
        type: array
      id:
        type: integer
      key:
        type: string
      location:
        type: string
      name:
        type: string
      targets:
        items:
          $ref: '#/definitions/model.Target'
        type: array
      updatedAt:
        type: string
    type: object
  model.Heartbeat:
    properties:
      checker:
        $ref: '#/definitions/model.Checker'
      checkerId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      latency:
        type: integer
      status:
        $ref: '#/definitions/model.TargetStatus'
      target:
        $ref: '#/definitions/model.Target'
      targetId:
        type: integer
      timestamp:
        type: string
      updatedAt:
        type: string
    type: object
  model.Incident:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      duration:
        $ref: '#/definitions/time.Duration'
      end:
        type: string
      id:
        type: integer
      ongoing:
        type: boolean
      start:
        type: string
      targetId:
        type: integer
      updatedAt:
        type: string
    type: object
  model.Notification:
    properties:
      alarms:
        items:
          $ref: '#/definitions/model.Alarm'
        type: array
      createdAt:
        type: string
      credentials:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      type:
        $ref: '#/definitions/model.NotificationType'
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  model.NotificationType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-varnames:
    - Discord
    - Slack
    - Pushbullet
    - Email
    - Gotify
    - GenericWebhook
...
  model.Session:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      expiresAt:
        type: string
      id:
        type: integer
      ipAddress:
        type: string
      lastActive:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userAgent:
        type: string
      userId:
        type: integer
    type: object
...
  model.Target:
    properties:
      alarms:
        items:
          $ref: '#/definitions/model.Alarm'
        type: array
      checkers:
        items:
          $ref: '#/definitions/model.Checker'
        type: array
      checksDown:
        type: integer
      checksUp:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      heartbeats:
        items:
          $ref: '#/definitions/model.Heartbeat'
        type: array
      httpInfo:
        $ref: '#/definitions/model.TargetHTTPInfo'
      id:
        type: integer
      incidents:
        items:
          $ref: '#/definitions/model.Incident'
        type: array
      lastCheck:
        type: string
      lastStatus:
        $ref: '#/definitions/model.TargetStatus'
      maxRetries:
        type: integer
      name:
        type: string
      paused:
        type: boolean
      pingInfo:
        $ref: '#/definitions/model.TargetPingInfo'
      timeout:
        type: integer
      type:
        $ref: '#/definitions/model.TargetType'
      updatedAt:
        type: string
      usedRetries:
        type: integer
      userId:
        type: integer
    type: object
  model.TargetHTTPInfo:
    properties:
      allowedCodes:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      followRedirects:
        type: boolean
      host:
        type: string
      id:
        type: integer
      targetId:
        type: integer
      updatedAt:
        type: string
      verifySSLCert:
        type: boolean
    type: object
  model.TargetPingInfo:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      host:
        type: string
      id:
        type: integer
      targetId:
        type: integer
      updatedAt:
        type: string
    type: object
  model.TargetStatus:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Up
    - Down
    - Unknown
  model.TargetType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      targetTypeMax: '"sentinel" value'
    x-enum-descriptions:
    - ""
    - ""
    - '"sentinel" value'
    x-enum-varnames:
    - PING
    - HTTP
    - targetTypeMax
  model.User:
    properties:
      admin:
        type: boolean
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      enabled:
        type: boolean
      id:
        type: integer
      notifications:
        items:
          $ref: '#/definitions/model.Notification'
        type: array
      sessions:
        items:
          $ref: '#/definitions/model.Session'
        type: array
      targets:
        items:
          $ref: '#/definitions/model.Target'
        type: array
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  notification.UpdateNotificationRequest:
    properties:
      credentials:
        description: Credentials must be valid for the selected type, e.g. webhook
          URLs have to start with http:// or https://, if supplied
        type: string
      name:
        description: Name must not be blank if supplied
        type: string
      type:
        allOf:
        - $ref: '#/definitions/model.NotificationType'
        description: Type must be a number in range of 0-5 (inclusive), if supplied.
          If type is changed, credentials valid for the new type also have to be supplied.
    type: object
...
  notification.getNotificationSuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/model.Notification'
    type: object
...
  target.CreateTargetRequest:
    properties:
      checkerIDs:
        description: At least one checker must be supplied, all checkers must exist
        items:
          type: integer
        minItems: 1
        type: array
      httpInfo:
        allOf:
        - $ref: '#/definitions/target.HTTPInfoCreateRequest'
        description: Must be supplied if type is 1 (HTTP)
      maxRetries:
        description: Max retries is required
        type: integer
      name:
        description: Name must not be blank
        type: string
      pingInfo:
        allOf:
        - $ref: '#/definitions/target.PingInfoCreateRequest'
        description: Must be supplied if type is 0 (PING)
      timeout:
        description: Timeout is required
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/model.TargetType'
        description: Type must be 0 (PING) or 1 (HTTP)
    required:
    - checkerIDs
    - name
    type: object
  target.HTTPInfoCreateRequest:
    properties:
      allowedCodes:
        description: HTTP response codes separated by a space. Must contain at least
          one code. All codes must be exactly 3 digits long.
        type: string
      followRedirects:
        type: boolean
      host:
        description: Host must start with http:// or https://
        type: string
      verifySSLCert:
        type: boolean
    required:
    - allowedCodes
    - followRedirects
    - host
    - verifySSLCert
    type: object
  target.HTTPInfoUpdateRequest:
    properties:
      allowedCodes:
        description: HTTP response codes separated by a space. Must contain at least
          one code if supplied. All codes must be exactly 3 digits long.
        type: string
      followRedirects:
        type: boolean
      host:
        description: Host must start with http:// or https:// if supplied
        type: string
      verifySSLCert:
        type: boolean
    type: object
  target.PingInfoCreateRequest:
    properties:
      host:
        description: Host must not be blank
        type: string
    required:
    - host
    type: object
  target.PingInfoUpdateRequest:
    properties:
      host:
        description: Host must not be blank if pingInfo is supplied in UpdateTargetRequest
        type: string
    type: object
  target.UpdateTargetRequest:
    properties:
      checkerIDs:
        description: Must contain at least one checker ID, all checkers must exist.
          This is a "replace update"
        items:
          type: integer
        type: array
      httpInfo:
        $ref: '#/definitions/target.HTTPInfoUpdateRequest'
      maxRetries:
        type: integer
      name:
        description: Name must not be blank if supplied
        type: string
      pingInfo:
        $ref: '#/definitions/target.PingInfoUpdateRequest'
      timeout:
        type: integer
    type: object
  target.createTargetSuccessResponse:
    properties:
      code:
        example: 201
        type: integer
      data:
        $ref: '#/definitions/model.Target'
    type: object
  target.getTargetSuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/model.Target'
    type: object
  target.listTargetsSuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data:
        items:
          $ref: '#/definitions/model.Target'
        type: array
    type: object
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    format: int64
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
  user.ChangePasswordRequest:
    properties:
      newPassword:
        description: New password must contain at least 1 uppercase letter, 1 lowercase
          letter, 1 number and must consist of at least 8 characters.
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  user.getUserSuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/model.User'
    type: object
host: localhost:8080
info:
  contact: {}
  description: API docs for bmonitord v3
  title: bmonitord v3 API
  version: "1.0"
paths:
...
  /notification/:id:
    delete:
      consumes:
      - application/json
      description: Allows a user to delete a notification with specified ID
      parameters:
      - description: ID of notification to be deleted
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/helpers.GenericDeleteSuccessResponse'
        "400":
          description: Returned when given ID couldn't be parsed.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "401":
          description: Returned when user sending the request supplies an invalid
            auth token.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "403":
          description: Returned when account of user sending the request is disabled.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "404":
          description: Returned when a notification with given ID couldn't be found.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "500":
          description: Returned when a DB interaction fails.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a notification by ID
      tags:
      - notification
    get:
      consumes:
      - application/json
      description: Allows a user to retrieve information about notification with specified
        ID
      parameters:
      - description: ID of notification to get
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.getNotificationSuccessResponse'
        "400":
          description: Returned when given ID couldn't be parsed.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "401":
          description: Returned when user sending the request supplies an invalid
            auth token.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "403":
          description: Returned when account of user sending the request is disabled.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "404":
          description: Returned when a notification with given ID couldn't be found.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "500":
          description: Returned when a DB interaction fails.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a notification by ID
      tags:
      - notification
    patch:
      consumes:
      - application/json
      description: Allows a user to update a notification with specified ID
      parameters:
      - description: ID of notification to update
        in: path
        name: id
        required: true
        type: integer
      - description: Changes to make to the notification
        in: body
        name: updateReq
        required: true
        schema:
          $ref: '#/definitions/notification.UpdateNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.getNotificationSuccessResponse'
        "400":
          description: Returned when given ID couldn't be parsed or request body doesn't
            match requirements.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "401":
          description: Returned when user sending the request supplies an invalid
            auth token.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "403":
          description: Returned when account of user sending the request is disabled.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "404":
          description: Returned when a notification with given ID couldn't be found.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "500":
          description: Returned when a DB interaction fails.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
      security:
      - BearerAuth: []
      summary: Update notification
      tags:
      - notification
...
  /notification/:targetID:
    delete:
      consumes:
      - application/json
      description: Allows a user to delete a target with specified ID
      parameters:
      - description: ID of target to delete
        in: path
        name: targetID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/helpers.GenericDeleteSuccessResponse'
        "400":
          description: Returned when given ID couldn't be parsed.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "401":
          description: Returned when user sending the request supplies an invalid
            auth token.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "403":
          description: Returned when account of user sending the request is disabled.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "404":
          description: Returned when a target with given ID couldn't be found.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "500":
          description: Returned when a DB interaction fails.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete target by ID
      tags:
      - target
...
  /target/:
    get:
      consumes:
      - application/json
      description: Allows a user to get a list of their targets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/target.listTargetsSuccessResponse'
        "401":
          description: Returned when user sending the request supplies an invalid
            auth token.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "403":
          description: Returned when account of user sending the request is disabled.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "500":
          description: Returned when a DB interaction fails.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all targets
      tags:
      - target
    post:
      consumes:
      - application/json
      description: Allows a user to create a target
      parameters:
      - description: Information about new target
        in: body
        name: createReq
        required: true
        schema:
          $ref: '#/definitions/target.CreateTargetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/target.createTargetSuccessResponse'
        "400":
          description: Returned when request body doesn't match requirements.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "401":
          description: Returned when user sending the request supplies an invalid
            auth token.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "403":
          description: Returned when account of user sending the request is disabled.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "500":
          description: Returned when a DB interaction fails.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a target
      tags:
      - target
  /target/:targetID:
    get:
      consumes:
      - application/json
      description: Allows a user to retrieve information about target with specified
        ID
      parameters:
      - description: ID of target to get
        in: path
        name: targetID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/target.getTargetSuccessResponse'
        "400":
          description: Returned when given ID couldn't be parsed.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "401":
          description: Returned when user sending the request supplies an invalid
            auth token.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "403":
          description: Returned when account of user sending the request is disabled.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "404":
          description: Returned when a target with given ID couldn't be found.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "500":
          description: Returned when a DB interaction fails.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
      security:
      - BearerAuth: []
      summary: Get target by ID
      tags:
      - target
    patch:
      consumes:
      - application/json
      description: Allows a user to change information about a target
      parameters:
      - description: ID of target to update
        in: path
        name: targetID
        required: true
        type: integer
      - description: New information about target
        in: body
        name: updateReq
        required: true
        schema:
          $ref: '#/definitions/target.UpdateTargetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/target.getTargetSuccessResponse'
        "400":
          description: Returned when given ID couldn't be parsed or request body doesn't
            match the requirements.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "401":
          description: Returned when user sending the request supplies an invalid
            auth token.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "403":
          description: Returned when account of user sending the request is disabled.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "404":
          description: Returned when a target with given ID couldn't be found.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "500":
          description: Returned when a DB interaction fails.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
      security:
      - BearerAuth: []
      summary: Update target
      tags:
      - target
...
  /user/:
    delete:
      consumes:
      - application/json
      description: Allows user to delete their own account and all resources they
        own
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/helpers.GenericDeleteSuccessResponse'
        "401":
          description: Returned when user sending the request supplies an invalid
            auth token.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "403":
          description: Returned when account of user sending the request is disabled.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "500":
          description: Returned when a DB interaction fails.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete current user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Allows user to get information about their own account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.getUserSuccessResponse'
        "401":
          description: Returned when user sending the request supplies an invalid
            auth token.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "403":
          description: Returned when account of user sending the request is disabled.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "500":
          description: Returned when a DB interaction fails.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - user
  /user/password:
    patch:
      consumes:
      - application/json
      description: Allows user to change their own password
      parameters:
      - description: Old and new password
        in: body
        name: changePasswordReq
        required: true
        schema:
          $ref: '#/definitions/user.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.getUserSuccessResponse'
        "400":
          description: Returned when request body is malformed or new password is
            too weak.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "401":
          description: Returned when user sending the request supplies an invalid
            auth token.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "403":
          description: Returned when account of user sending the request is disabled
            or provided old password doesn't match.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
        "500":
          description: Returned when a DB interaction or password hashing fails.
          schema:
            $ref: '#/definitions/helpers.GenericErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - user
...
securityDefinitions:
  BearerAuth:
    description: Header value has to start with "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"